+++ tests/testthat.R
@@ -1,0 +1,4 @@
+library(testthat)
+library(CAVAanalytics)
+
+test_check("CAVAanalytics")